+++ Changelog.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [UNRELEASED]
+### Added
+ - Content property to HTML and Content components Logic.
+
 ## 3.6.11
 ### Changed
  - Export the "moment" object in FormioUtils so that it can be accessed from outside libraries.